# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Goyais Contracts
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8787
paths:
...
  /v1/workspaces:
    get:
      summary: List workspaces
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Workspace list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceListEnvelope'
    post:
      summary: Create remote workspace metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceRequest'
      responses:
        '201':
          description: Created workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/StandardErrorResponse'
...
components:
  parameters:
    CursorParam:
      in: query
      name: cursor
      schema:
        type: string
      description: Opaque cursor; current implementation uses numeric offsets.
    LimitParam:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Page size for cursor-based pagination.
...
  responses:
    StandardErrorResponse:
      description: Standard API error body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardError'
  schemas:
...
    Workspace:
      type: object
      required:
        - id
        - name
        - mode
        - hub_url
        - is_default_local
        - created_at
        - login_disabled
        - auth_mode
      properties:
        id:
          type: string
        name:
          type: string
        mode:
          type: string
          enum: [local, remote]
        hub_url:
          type: [string, 'null']
        is_default_local:
          type: boolean
        created_at:
          type: string
          format: date-time
        login_disabled:
          type: boolean
        auth_mode:
          type: string
          enum: [disabled, password_or_token, token_only]
...
    WorkspaceListEnvelope:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Workspace'
        next_cursor:
          type: [string, 'null']
    CreateWorkspaceRequest:
      type: object
      required: [hub_url]
      properties:
        name:
          type: string
        hub_url:
          type: string
        login_disabled:
          type: boolean
          default: false
        auth_mode:
          type: string
          enum: [password_or_token, token_only]
          default: password_or_token
...
    StandardError:
      type: object
      required: [code, message, details, trace_id]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        trace_id:
          type: string
...